Add flow.SignUserMessage function
