+++ sign.go
@@ -1,0 +1,54 @@
+/*
+ * Flow Go SDK
+ *
+ * Copyright 2019-2020 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package flow
+
+import (
+	"fmt"
+
+	"github.com/onflow/flow-go-sdk/crypto"
+)
+
+const domainTagLength = 32
+
+// TransactionDomainTag is the prefix of all signed transaction payloads.
+//
+// A domain tag is encoded as UTF-8 bytes, right padded to a total length of 32 bytes.
+var TransactionDomainTag = paddedDomainTag("FLOW-V0.0-transaction")
+
+// UserDomainTag is the prefix of all signed user space payloads.
+//
+// A domain tag is encoded as UTF-8 bytes, right padded to a total length of 32 bytes.
+var UserDomainTag = paddedDomainTag("FLOW-V0.0-user")
+
+func paddedDomainTag(s string) [domainTagLength]byte {
+	var tag [domainTagLength]byte
+
+	if len(s) > domainTagLength {
+		panic(fmt.Sprintf("domain tag %s cannot be longer than %d characters", s, domainTagLength))
+	}
+
+	copy(tag[:], s)
+
+	return tag
+}
+
+func SignUserMessage(signer crypto.Signer, message []byte) ([]byte, error) {
+	message = append(UserDomainTag[:], message...)
+	return signer.Sign(message)
+}
